### 2025-02-20 #7820 @editorialbot
[PRE REVIEW]: FRAGMENT-MNP: A model of micro- and nanoplastic fragmentation in the environment
**Submitting author:** @samharrison7 (<a href="http://orcid.org/0000-0001-8491-4720">Sam Harrison</a>)
**Repository:** https://github.com/microplastics-cluster/fragment-mnp
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.3
