### 2025-02-20 #7820 @samharrison7
[Michael C. Denes](https://reviewers.joss.theoj.org/reviewers/4420) might be a good reviewer, based on their topics of interest (including plastic pollution) and languages (Python).
