### 2025-03-07 27a63ae
Moving a few dependencies to test env
