+++ paper/paper.md
@@ -58,7 +58,7 @@ affiliations:
   - name: School of Earth and Environment, University of Leeds, Leeds, LS2 9JT, UK
     index: 8
 date: 17 February 2025
-bibliography: paper.bib
+bibliography: paper/paper.bib
 ---
 
 # Summary
@@ -99,7 +99,7 @@ output = fmnp.run()
 output.plot()
 ```
 
-![Example model output showing the time evolution of particle size distributions undergoing fragmentation](./fragmentmnp_example.png){height="900px"}
+![Example model output showing the time evolution of particle size distributions undergoing fragmentation](./paper/fragmentmnp_example.png){height="900px"}
 
 # Related work
 
