+++ pyproject.toml
@@ -16,9 +16,6 @@ schema = "^0.7.5"
 plotly = "^5.8.0"
 pandas = "^1.4.2"
 Pillow = ">=9.1.1"
-ipykernel = "^6.16.2"
-jupyter-core = ">=4.11.2"
-sphinxcontrib-mermaid = "^1.0.0"
 
 [tool.poetry.group.dev.dependencies]
 flake8 = "^4.0.1"
@@ -27,7 +24,9 @@ pytest-cov = "^3.0.0"
 ipykernel = "^6.13.0"
 ipywidgets = "^7.7.0"
 jupyter-book = "^1.0.3"
+jupyter-core = ">=4.11.2"
 notebook = ">=6.4.12"
+sphinxcontrib-mermaid = "^1.0.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
