### 2025-03-07 ba53915
Moving some dependencies to dev env
