### 2025-04-06 #7820 @Kevin-Mattheus-Moerman
@espottesmith do you think this paper is something you could help edit (provided it is not too far off topic)? Thanks
