### 2025-04-06 #7820 @espottesmith
Yep, I can take this one on!
