### 2025-04-06 #7820 @espottesmith
Hi @ashahide @nanhung @flor14 @simonom,

Considering your field(s) of expertise and experience with JOSS, would any of you be willing to review this submission? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.07820/joss.07820/10.21105.joss.07820.pdf
Software: https://github.com/microplastics-cluster/fragment-mnp
