### 2025-04-09 #7820 @flor14
Hello @espottesmith.
Thank you for the invitation! I really enjoy the process of reviewing JOSS articles and software. However, I’m unable to commit to a review in the short term, so I’ll have to decline this time.
