### 2025-04-10 #7820 @joeweaver
> Hey [@puruckertom](https://github.com/puruckertom) [@joeweaver](https://github.com/joeweaver), any chance y'all could contribute a review to this package/paper?

It looks interesting - in a workshop right now, but I'll be able to give a definite response by noon tomorrow (probably earlier).
