### 2025-04-11 #7820 @joeweaver
I can commit to getting this reviewed within the next three weeks, probably faster.
