### 2025-04-16 18d34e5
(Trying to) fix paper generation issue
