### 2025-04-18 #8061 @joeweaver
I'm running into a bit of an odd issue with the example code.  I suspect it's more a problem with my environment than the software.

When run as a normal script the figure is not displayed. This occurs either when running from the command line or within PyCharm.

When run in debug mode within PyCharm, the figure will display,  so long as breakpoints are set in the correct place.  When I get to the end of the script, the displayed plot is closed.  I wondered if during a non-debug run the plot was showing briefly and closing, but there were no indications of this, not even a very brief flicker.  Adding a call to `sleep()` just before exit doesn't seem to make a difference.

I can use the returned values  from the call to `plot` to display the figure, but they also disappear on program exit.

No issues in other code on my machine using matplotlib, but I am in a fresh environment for this program (matplotlib for this env is 3.8.0 and this is the same as in some of the other envs where I have no display issue).

I'm using Ubuntu 20.04 and i3wm as my window manager (this may be part of the issue, but has not been problematic before).

So far, everything else looks like it's working as reported and I'm happy to keep reviewing using workarounds to view the plots.  Unless everyone else has the same issue, and if I can't figure out the problem, I'm happy to just list it as an issue in the repo but don't think it should block publication acceptance.
