### 2025-04-19 #8061 @samharrison7
Thanks for flagging this issue! I'm away from keyboard at the moment so can't test it, but my first instinct is that this is because you're running a normal script and not something interative like a Jupyter Notebook. The code doesn't have a `plt.show()` call in there anywhere, which isn't required in notebooks but is for scripts to show plots. As a test, can you try the following?

```python
import matplotlib.pyplot as plt
from fragmentmnp import FragmentMNP
from fragmentmnp.examples import minimal_config, minimal_data

# Create the model, pass it example config and data, then run it
fmnp = FragmentMNP(minimal_config, minimal_data)
output = fmnp.run()
# Plot the mass concentration time series
output.plot()
plt.show()
```

If we confirm this is the problem, I'll do some thinking about how best to solve it (i.e. updating the docs or adding the `plt.show()` calls to the package).
