### 2025-04-19 #8061 @joeweaver
You're absolutely right, an explicit call to `plt.show()`  fixes it.

My suggestion would be to add something to the docs either indicating the example is to be run inside a notebook (perhaps with a Note: callout) or defaulting to have plt.show().  Depends on how you view the main audience using FRAGMENT-MNP.
