+++ .github/workflows/gh-pages.yml
@@ -55,3 +55,4 @@ jobs:
       with:
         github_token: ${{ secrets.GITHUB_TOKEN }}
         publish_dir: ./docs/_build/html
+        cname: fragmentmnp.ceh.ac.uk
