### 2025-04-22 #8061 @samharrison7
Thanks for confirming. I've gone for a hybrid option: I added a `show=True|False` parameter to the `plot()` function, which if `True` calls `plt.show()` in the function. It defaults to `False` as I suspect the main audience will be using it via notebooks, and calling `plt.show()` in `plot()` means you can't make use of the `fix, ax` passed back via the `plot()` function ([e.g.](https://fragmentmnp.ceh.ac.uk/plotting.html#advanced-control-through-matplotlib)). I've documented this in a few places, including the [quickstart](https://fragmentmnp.ceh.ac.uk/intro.html). This is all in a new release (1.1.5).
