### 2025-04-29 #8061 @joeweaver
I can confirm that fix works, and is elegant to boot.  Easter holidays have ended and I intend to finish up comments by the next Monday, probably sooner.  Informally, I haven't seen any big issues so I anticipate the review will be mainly suggestions for some quick clarifications.
