### 2025-05-02 #8061 @joeweaver
`FRAGMENT-MNP` is a worthwhile advance in modelling plastic fragmentation; specifically how particle size distributions can change over time.  It is build on solid principles and the underlying code is consistent and professional. With a minor exception, already handled, it worked right out of the box and performed as expected. I am not an expert in plastic fragmentation, but I do have some experience in particle size distributions and the software raises no red flags regarding correctness.

There are no major issues, and I thank the authors for the attention to quality making my task as a reviewer easy. There are a few minor issues to address for some check boxes (mostly technicalities).  I will also signpost any minor suggestions offered which may simply be a matter of style and which are not related to whether or not this paper should be accepted.

### Documentation

- **Statement of Need:** Neither the Github repo nor the docs have a statement of need.  That statement is expressed well in the paper. Adding that statement (or a paraphrase if self-plagiarism is a concern,  @espottesmith what's JOSS policy on that? ) to at least the documentation site would satisfy this checkmark as far as I'm concerned. I'd also suggest adding similar to the repo `README`, since it may be the first place people find `FRAGMENT-MNP` and it's not immediately clear from that text why `FRAGMENT-MNP` exists.

- **Install instructions:** Installation is straightforward via `pip` and all the requirements *are* in the `toml` file.  As a quality of life suggestion, adding the minimum Python version to the docs/repo would have saved me a very minor bit of work.

- **Community Guidelines:** This passes already, all information is in the documentation or is implicit as part of the usual git/Github practices (e.g., issue tracking). If the authors wish, they might want to play with using the criteria from the checklist item as explicit section names in the documentation to make that sort of info a bit easier for novice devs.

### Software Paper

The paper itself is fine.  There is one minor issue with the summary.  Given that the section is intended is for non-specialist audiences, the phrase 'time evolution of particle size distributions' could be a simplified a bit. That phrase actually read fine to me at first until I tried to read it as someone who's not dealt with time varying PSDs before.  I'd suggest simplifying the language further - perhaps  'how pieces of plastic can breakup and aggregate over time' if it still reflects the author's intentions? Given that the goal is to write for a non-*specialist* audience but non necessarily a non-*technical* one, I'm also open to some middle-ground phrasing.
