### 2025-05-19 #8061 @ashahide
FRAGMENT-MNP is a carefully and professionally designed toolbox for modeling the kinetics of plastic degradation and fragmentation. The authors have clearly devoted much time and effort to practical software development, packing, and documentation, making both the usage and review very straight forward. While this is not my area of expertise, I have a background in chemical engineering and the modeling details seem reasonable.

**Comments to Address**
Unless I am mistaken, a statement of need should be added to both the paper and the documentation to meet the JOSS requirements. Additionally, a short statement in the paper (perhaps in the "Related Work" section?) mentioning alternative software (or the lack of alternatives, highlighting need) would also strengthen the paper.

I have no additional concerns. Both the paper and software are excellent and I believe they are ready for publication pending the minor points above. I’ve included some general, non-blocking suggestions below.

**General Thoughts**
Much of the code in fragmentmnp.py could be further modularized. While the current implementation is clean and well-documented, many individual functions handle a lot of logic. Adding smaller, single-purpose functions would improve:

1.     Readability
2.     Ease of testing
3.     Maintainability and future development

For instance, lines 403–455 could be factored out as an assign_params() helper function, reducing verbosity and enabling unit testing. A similar opportunity exists for the logic in lines 71–108.

I may have missed it, but if complete regression tests or end-to-end examples are not currently part of the test suite, they would be a welcome addition.

Is there a use case for running simulations across parameter ranges? For example, grid search, Monte Carlo simulations, or sensitivity analyses? This is something I’d personally find interesting, though I understand it may be beyond the scope or goals of the current release.
