### 2025-05-21 microplastics-cluster/fragment-mnp#26 @samharrison7
Bootstrapping, sensitivity/uncertainty, calibration, etc
Often the model is run iteratively across e.g. a range of parameters as part of a sensitivity/uncertainty analysis, or as part of calibrating model parameters to measurement data. It would be nice to build this functionality into the model, though it requires some thinking about how best to do it.
