### 2025-05-21 microplastics-cluster/fragment-mnp#27 @samharrison7
Better modularity in `fragmentmnp.py`
This file is becoming a bit unweildy. I made a couple of changes based on review comments (https://github.com/openjournals/joss-reviews/issues/8061), but it could do with a full go-through to see what else could be done. E.g. it might be a good idea to split out the static methods into separate utility file(s).
