+++ paper/paper.md
@@ -63,7 +63,7 @@ bibliography: paper.bib
 
 # Summary
 
-The degradation and fragmentation of plastics in the environment is an important but relatively poorly characterised process. Fragmentation leads to the formation of micro- and nanoplastics, and research has shown that particle size and shape, and thereby fragmentation, impacts a variety of processes, such as the ability of organisms to uptake plastics, and the movement of plastics around the environment [@ThorntonHampton:2022; @Thompson:2024]. In this paper, we present the FRAGMENT-MNP model as an open-source mechanistic model of micro- and nanoplastic degradation and fragmentation. FRAGMENT-MNP predicts the time evolution of particle size distributions, providing invaluable insights into fragmentation dynamics to help develop our understanding of plastic degradation and fragmentation, and offering predictive capabilities to enable better prediction of the fate and exposure of plastics in the environment.
+The degradation and fragmentation of plastics in the environment is an important but relatively poorly characterised process. Fragmentation leads to the formation of micro- and nanoplastics, and research has shown that particle size and shape, and thereby fragmentation, impacts a variety of processes, such as the ability of organisms to uptake plastics, and the movement of plastics around the environment [@ThorntonHampton:2022; @Thompson:2024]. In this paper, we present the FRAGMENT-MNP model as an open-source mechanistic model of micro- and nanoplastic degradation and fragmentation. FRAGMENT-MNP predicts how pieces of plastic can break part (fragment), providing invaluable insights into fragmentation dynamics to help develop our understanding of plastic degradation and fragmentation, and offering predictive capabilities to enable better prediction of the fate and exposure of plastics in the environment.
 
 # Statement of need
 
