+++ src/fragmentmnp/fragmentmnp.py
@@ -1,4 +1,4 @@
-from typing import Tuple
+from typing import Tuple, Sequence
 import numpy as np
 import numpy.typing as npt
 from scipy.integrate import solve_ivp
@@ -64,50 +64,9 @@ class FragmentMNP():
         self.k_frag = np.empty((self.n_size_classes, self.n_timesteps))
         self.k_diss = np.empty((self.n_size_classes, self.n_timesteps))
         self.k_min = np.empty((self.n_timesteps,))
-        # Calculate the rate constant distributions. If we've been given
-        # a dict in data, use the contained params to create the distribution.
-        # Else, presume that we've been given a scalar (validation will make
-        # sure this is so) and use that as the average.
+        # Calculate the rate constant distributions
         for k in ['k_frag', 'k_diss', 'k_min']:
-            if isinstance(data[k], dict):
-                k_f = data[k]['k_f']
-                k_0 = data[k]['k_0']
-                is_compound = data[k]['is_compound']
-                # Get the params from the dict, excluding the average
-                params = {n: p for n, p in data[k].items()
-                          if n not in ['k_f', 'k_0']}
-            else:
-                k_f = data[k]
-                k_0 = 0.0
-                is_compound = True
-                params = {}
-            # If k_frag or k_diss, we want to calculate a 2D (s, t) distribution,
-            # and if k_min, we just want a 1D (t) distribution
-            if k in ['k_frag', 'k_diss']:
-                k_dist = self.set_k_distribution(dims={'s': self.surface_areas,
-                                                       't': self.t_grid},
-                                                 k_f=k_f, k_0=k_0,
-                                                 params=params,
-                                                 is_compound=is_compound)
-                # If the rate constant is k_frag, then no fragmentation is
-                # allowed from the smallest size class and therefore we
-                # manually set this to zero
-                if k == 'k_frag':
-                    k_dist[0, :] = 0.0
-            else:
-                k_dist = self.set_k_distribution(dims={'t': self.t_grid},
-                                                 k_f=k_f, k_0=k_0,
-                                                 params=params,
-                                                 is_compound=is_compound)
-            # Check no values are less than zero
-            if np.any(k_dist < 0.0):
-                msg = (f'Value for {k} distribution calculated from input '
-                       'data resulted in negative values. Ensure '
-                       f'distribution params are such that all {k} values '
-                       'are positive.')
-                raise FMNPDistributionValueError(msg)
-
-            # Set self.k_[frag|diss|min] as this distribution
+            k_dist = self.set_rate_constant(data[k], k)
             setattr(self, k, k_dist)
 
     def run(self) -> FMNPOutput:
@@ -260,6 +219,74 @@ class FragmentMNP():
                              'neither were found.')
         return psd
 
+    def set_rate_constant(self, params: dict, name: str) \
+            -> npt.NDArray[np.float64]:
+        """
+        Set a rate constant either for fragmentation, dissolution or
+        mineralisation. The function largely parses input data and
+        deals with defaults etc, before calling `set_k_distribution`
+        to create the rate constant distribution.
+
+        Parameters
+        ----------
+        params : dict
+            The param dict from the model input data, from which the rate
+            constant parameters are extracted. See the notes in
+            `set_k_distribution` for more details on the expected format
+        name : str
+            The name of the rate constant to set. This should be one of
+            `k_frag`, `k_diss` or `k_min`
+
+        Returns
+        -------
+        np.ndarray
+            The rate constant distribution
+        """
+        # If we've been given a dict in data, use the contained params to
+        # create the distribution. Else, presume that we've been given a
[…]
