+++ CONTRIBUTING.md
@@ -2,19 +2,23 @@
 
 Thanks for considering contributing to FRAGMENT-MNP. We very much welcome your input.
 
-## Bug reports
-Whilst bug reports are very welcome, we encourage pull requests to accompany these. Reports should be made via the issue tracker and should include enough information to be fully reproducible. This includes information on the system and compiler you are using, as well as detailed information about what you were doing that caused the bug.
-
-## New and updated features
-We welcome new and updated features via pull requests. Only completed features that do not require any further coding to become functional will be accepted. If you have an idea for a feature but don't have the time to code it, then consider raising a feature request issue. If you have some code towards a feature but won't have time to finish it, then raise an issue highlighting this, and if we think the feature is valuable enough, we may open a new feature branch for you to submit your code to.
+## Contibuting to the model
+We welcome new and updated features via pull requests. Only completed features that do not require any further coding to become functional will be accepted. If you have an idea for a feature but don't have the time to code it, then consider raising a feature request issue. If you have some code towards a feature but won't have time to finish it, then raise an issue highlighting this, and if we think the feature is valuable enough, we may open a new feature branch for you to submit your code to. If you are unsure of the suitability of your feature, raise an issue to discuss it first.
 
 Pull requests with new or updated features should, alongside the updated code, include relevant tests (in the `tests/` directory) and documentation (in the `docs/` directory).
 
+
+## Reporting issues or problems with the model
+Whilst bug reports are very welcome, we encourage pull requests to accompany these. Reports should be made via the issue tracker and should include enough information to be fully reproducible. This includes information on the system and, if applicable, environment manager you are using, as well as detailed information about what you were doing that caused the issue.
+
+## Seeking support
+If you would like general support on using the model, for example if you would like guidance applying the model to your own data, then get in touch with Sam Harrison at sharrison@ceh.ac.uk.
+
 ## Which branch?
-Please send pull requests for all contributions (bug reports and features) to the develop branch.
+Please send pull requests for all contributions (issue reports and features) to the develop branch.
 
 ## Security vulnerabilities
-If you discover a security vulnerability, please send an email to Sam Harrison at samharrison.xg@gmail.com. Please do not submit an issue with the vulnerability.
+If you discover a security vulnerability, please send an email to Sam Harrison at sharrison@ceh.ac.uk. Please do not submit an issue with the vulnerability.
 
 ## Coding style
 Please use PEP8 and comment your code well. We are using [NumPy/SciPy docstrings](https://numpydoc.readthedocs.io/en/latest/format.html) in this project. Remember that code is (probably) read more often than it is written.
