+++ CHANGELOG.md
@@ -54,7 +54,8 @@ Notable changes to FRAGMENT-MNP will be documented here. We are using [semantic 
 * Initial model version.
 
 
-[Unreleased]: https://github.com/microplastics-cluster/fragment-mnp/compare/1.1.5...HEAD
+[Unreleased]: https://github.com/microplastics-cluster/fragment-mnp/compare/1.1.6...HEAD
+[1.1.6]: https://github.com/microplastics-cluster/fragment-mnp/releases/tag/1.1.6
 [1.1.5]: https://github.com/microplastics-cluster/fragment-mnp/releases/tag/1.1.5
 [1.1.4]: https://github.com/microplastics-cluster/fragment-mnp/releases/tag/1.1.4
 [1.1.3]: https://github.com/microplastics-cluster/fragment-mnp/releases/tag/1.1.3
