### 2025-05-21 microplastics-cluster/fragment-mnp#28 @samharrison7
Better contributing docs
The docs are laking a bit in the detail about how to contribute / be part of the community at the moment, and could do a much better job of laying out how people can get involved (as referenced in the JOSS review: https://github.com/openjournals/joss-reviews/issues/8061). This is currently only covered in CONTRIBUTING.md. We should restructure a little to at least address:
* Contributing to the model
* Reporting issues or problems with the model
* Seeking support
