### 2025-05-21 #8061 @samharrison7
Thanks so much for both of the reviews, we really appreciate the time and effort 🙂 Thank you also for the positive comments about the overall model, this is nice to hear.

I'll comment on the issues identified here, and any changes that I've made are in a new release (https://github.com/microplastics-cluster/fragment-mnp/releases/tag/1.1.6)

Comments from @joeweaver:
- **Statement of need:** I have now added a statement of need to the README and [documentation](https://fragmentmnp.ceh.ac.uk/intro.html#statement-of-need).
- **Minimum Python version:** I have added in the README and documentation a minimum Python version.
- **Community guidelines:** I agree that the documentation could do with a little work to make it more contribution-welcome. For the moment, I have tweaked the CONTRIBUTING.md file slightly to better follow the structure in the review checklist, which I think makes it clearer. I've also opened up an issue (https://github.com/microplastics-cluster/fragment-mnp/issues/28) where we will tackle this more thoroughly in the future.
- **Non-specialist language** in the paper: I completely agree that "time evolution of particle size distributions" is too specialist for this summary, and I like your suggested alternative. I have edited the paper accordingly.

Comments from @ashahide:
- **Statement of need:** Addressed above - the documentation and paper (and README) now have a statement of need. This statement of need already references the work that is most related to ours (e.g Quik et al 2023, Domercq et al 2023) in framing why we wanted to develop this model, including highlighting that there is, to our knowledge, no similar model available. If you don't think this covers the related work aspect explicitly enough, let me know and we can create a separate section.
- **Better modularity:** I agree, we could be better in splitting up the code in `fragmentmnp.py`. I have made a start on this with the two suggestions you made (thanks!) and hopefully this has helped. We will carry on working on this for future releases and I've opened up an issue to make sure we don't forgot it (https://github.com/microplastics-cluster/fragment-mnp/issues/27).
- **Regression tests:** We have a few tests in `tests/test_model.py` where we test the entire model, which acts as regression tests for us. E.g. https://github.com/microplastics-cluster/fragment-mnp/blob/8cf3a83f3f7f23ab27a46f441b23c2519587bbce/tests/test_model.py#L24-L43
- **Running simulations across parameter ranges:** This is a great point and something we do fairly often. We've already been thinking about the best way of implementing this in the model itself (it's quite tricky to get it right) and I think this is too big a task for this release, but I've opened up an issue to make sure this gets tackled in the future (https://github.com/microplastics-cluster/fragment-mnp/issues/26) - if for no other reason than it would be useful for us too!

I hope that has addressed everything, but do say if I've missed anything or if you think I haven't addressed something fully enough.

Thanks once again for you time 🙂
