### 2025-05-29 #8061 @samharrison7
I've just archived the latest release* (1.1.7) on Zenodo: https://doi.org/10.5281/zenodo.15543578. I also double checked all the affiliations etc, as requested.

*1.1.7 is exactly the same as 1.1.6 that I mentioned above, but I had to create a new release to trigger the automatic Zenodo upload... or at least I couldn't find an easy way around this!
