### 2025-05-29 f106595
Review updates to JOSS paper
