+++ paper/paper.bib
@@ -119,3 +119,38 @@
   doi = {10.1016/j.envsoft.2021.105291},
   keywords = {Environmental fate modelling,Freshwater,Microplastics,Open source,Plastic pollution},
 }
+
+@article{Pfohl:2025,
+  title = {Environmental Degradation and Fragmentation of Microplastics: Dependence on Polymer Type, Humidity, UV Dose and Temperature},
+  author = {Pfohl, Patrizia and Santizo, Katherine and Sipe, Joana and Wiesner, Mark and Harrison, Sam and Svendsen, Claus and Wohlleben, Wendel},
+  year = {2025},
+  month = mar,
+  journal = {Microplastics and Nanoplastics},
+  volume = {5},
+  number = {1},
+  pages = {7},
+  issn = {2662-4966},
+  doi = {10.1186/s43591-025-00118-9},
+}
+
+@techreport{eftec:2025,
+	title = {Option {Appraisal} for {Intentionally} {Added} {Microplastics}},
+	url = {https://sciencesearch.defra.gov.uk/ProjectDetails?ProjectId=21802},
+	language = {en},
+	number = {CB04121},
+	urldate = {2025-05-29},
+	institution = {Defra},
+	author = {{eftec} and {UKCEH}},
+	month = jan,
+	year = {2025},
+}
+
+
+@misc{Domercq:2025,
+	title = {{UTOPIA}},
+	url = {https://github.com/microplastics-cluster/UTOPIA_model},
+	urldate = {2025-05-29},
+	author = {Domercq, Prado and Praetorius, Antonia and Harrison, Sam and MacLeod, Matthew},
+	month = may,
+	year = {2025},
+}
