### 2025-05-29 #8061 @espottesmith
Alright, that all makes sense to me. Thank you for your additions to the paper, @samharrison7, and for your explanations. I'm now happy to recommend this manuscript (and code) for acceptance in JOSS!
