### 2025-05-29 4a74120
Fixing paper image path
