### 2025-05-29 #8061 @samharrison7
Whoops, we've got a missing image in the PDF! I've just pushed a fix (the path to the image was wrong).
