### 2025-05-29 #8061 @samharrison7
That's fixed the issue! I don't know how to trigger the final proof to be re-generated though.
