### 2025-05-30 #8061 @espottesmith
Wait, what's the issue, @samharrison7? When I looked, I didn't see a missing image. There's only one figure, yes?
