### 2025-05-30 #8061 @samharrison7
Yep there's only one figure. It wasn't showing for me on the "Final proof" link above, but now I changed the path to the image and have re-run editorialbot's generate pdf, the image is showing as it should.
