### 2025-06-07 #8061 @Kevin-Mattheus-Moerman
- [ ] Please spell out `UK` as `United Kingdom` in your affiliations (you can keep it as UK in `UK Centre for Ecology & Hydrology` as that is an institute name but please spell out country names).
- [ ] Check `goverment` which should read `government`
- [ ] Check `Iniative` which should read `Initiative`
